Clean up of some warnings
Most of MY errors

+++ EnemyAIMeleeBigBehavior.c
@@ -98,7 +98,7 @@ void EnemyAI_MeleeBig_AttackUpdate(GameObject* enemy, EnemyContainer* enemyConta
 
   if (enemyContainer->attackWindup > 0)
   {
-    enemyContainer->attackWindup -= AEFrameRateControllerGetFrameTime();
+    enemyContainer->attackWindup -= (float)AEFrameRateControllerGetFrameTime();
   }
 
   if (enemyContainer->attackWindup <= 0)
@@ -134,7 +134,7 @@ void EnemyAI_MeleeBig_CooldownUpdate(GameObject* enemy, EnemyContainer* enemyCon
   enemy->physics->velocity.y = 0;
   if (enemyContainer->attackCooldown > 0)
   {
-    enemyContainer->attackCooldown -= AEFrameRateControllerGetFrameTime();
+    enemyContainer->attackCooldown -= (float)AEFrameRateControllerGetFrameTime();
   }
 
   float distanceToPlayer = Vector2DSquareDistance(&(enemy->physics->position), &(enemy->target->physics->position));
